Task6 acceptance testing

+++ app/controllers/questions_controller.rb
@@ -1,10 +1,15 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
+
   expose :questions, -> { Question.all }
   expose :question
+  expose :answer, -> { Answer.new }
 
   def create
+    @exposed_question = current_user.author_questions.new(question_params)
+
     if question.save
-      redirect_to question_path(question)
+      redirect_to question_path(question), notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -19,8 +24,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Your question was successfully deleted.'
+    else
+      redirect_to questions_path, notice: 'You have no rights to do this.'
+    end
   end
 
   private
